Added react-router-dom package to project. Cleaned up/minor additions to server index.js. Trying out react router for different pages. Added preliminary components and pages for future development. Pushing react-router-dom changes and verify that it will work on Heroku.
