+++ client/src/pages/AboutUs.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+// Container for About Us page view
+const AboutUsPage = () => {
+  return(
+    <div className="AboutUsPage">
+      <p>Put About Us info here</p>
+    </div>
+  );
+}
+
+export default AboutUsPage;
