+++ client/src/pages/ExpensePage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+// Container for Expense page view
+const ExpensePage = () => {
+  return(
+    <div className="ExpensePage">
+      <p>Put expense data here</p>
+      <p>Probably put <code>ExpenseList.jsx</code> here</p>
+    </div>
+  );
+}
+
+export default ExpensePage;
