+++ client/src/pages/LandingPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+// Container for Landing Page view
+const LandingPage = () => {
+  return(
+    <div className="LandingPage">
+      <p>Put Landing Page info here</p>
+      <p>Maybe add a cool RV travel pic</p>
+    </div>
+  );
+}
+
+export default LandingPage;
